Adding header with functionality for walking elements.
